@@ -16,5 +16,5 @@ public class CityDto
     }
 
     public ICollection<PointOfInterestDto> PointsOfInterest { get; set; }
-        = new List<PointOfInterestDto>();
+        = [];
 }
