@@ -17,8 +17,8 @@ public class CitiesDataStore
                  Id = 1,
                  Name = "New York City",
                  Description = "The one with that big park.",
-                 PointsOfInterest = new List<PointOfInterestDto>()
-                 {
+                 PointsOfInterest =
+                 [
                      new() {
                          Id = 1,
                          Name = "Central Park",
@@ -27,15 +27,15 @@ public class CitiesDataStore
                          Id = 2,
                          Name = "Empire State Building",
                          Description = "A 102-story skyscraper located in Midtown Manhattan." },
-                 }
+                 ]
             },
             new()
             {
                 Id = 2,
                 Name = "Antwerp",
                 Description = "The one with the cathedral that was never really finished.",
-                PointsOfInterest = new List<PointOfInterestDto>()
-                 {
+                PointsOfInterest =
+                 [
                      new() {
                          Id = 3,
                          Name = "Cathedral of Our Lady",
@@ -44,15 +44,15 @@ public class CitiesDataStore
                          Id = 4,
                          Name = "Antwerp Central Station",
                          Description = "The the finest example of railway architecture in Belgium." },
-                 }
+                 ]
             },
             new()
             {
                 Id= 3,
                 Name = "Paris",
                 Description = "The one with that big tower.",
-                PointsOfInterest = new List<PointOfInterestDto>()
-                 {
+                PointsOfInterest =
+                 [
                      new() {
                          Id = 5,
                          Name = "Eiffel Tower",
@@ -61,7 +61,7 @@ public class CitiesDataStore
                          Id = 6,
                          Name = "The Louvre",
                          Description = "The world's largest museum." },
-                 }
+                 ]
             }
         ];
 
